requirements.txt
python-binance
numpy
urllib3==1.26.6
win10toast
python-dotenv